command: support on callbacks to protocol events-only
Every message or event that has a callback registered passes its
information through a generic callback interface, which doesn't care
about the source event. It always waits a message to be present, and
because of that it's impossible to register 'events-only' commands.

New field 'EventCode' on 'Message' structure allows callbacks to events
that doesn't require explicit message text, that just the name or code
of that event would be enough.